Fixed the removal of existing favicon tags
When the head contains many favicon tags (needed to support all variants
of modile and Windows 8 icons), the update of the live node list might be
fast enough to be done before the iteration is completed. This means that
some icons are not removed, which can break the resetting depending on the
browser. The list is now iterated backward so that updates don't change
the indexes of remaining elements.